@@ -12,30 +12,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.6",
-    "@angular/cdk": "^5.2.3",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.3",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@ngrx/effects": "^5.1.0",
-    "@ngrx/router-store": "^5.0.1",
-    "@ngrx/store": "^5.1.0",
+    "@angular/animations": "6.1.3",
+    "@angular/cdk": "^6.4.5",
+    "@angular/common": "6.1.3",
+    "@angular/compiler": "6.1.3",
+    "@angular/core": "6.1.3",
+    "@angular/forms": "6.1.3",
+    "@angular/http": "6.1.3",
+    "@angular/material": "^6.4.5",
+    "@angular/platform-browser": "6.1.3",
+    "@angular/platform-browser-dynamic": "6.1.3",
+    "@angular/router": "6.1.3",
+    "@ngrx/effects": "^6.0.0",
+    "@ngrx/router-store": "^6.0.0",
+    "@ngrx/store": "^6.0.0",
     "core-js": "^2.4.1",
     "material-design-icons": "^3.0.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.2.2",
     "wakanda-client": "git+https://git@github.com/wakanda/wakanda-javascript-client.git#feature/databrowser-requirements",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.1",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "6.1.4",
+    "@angular/compiler-cli": "6.1.3",
+    "@angular/language-service": "6.1.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -50,6 +51,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.9.2"
   }
 }
