@@ -1,15 +1,14 @@
 import { Component, OnInit, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/distinctUntilChanged';
+import { Observable } from 'rxjs';
+import { withLatestFrom, map } from 'rxjs/operators';
+
 import { Store, select } from '@ngrx/store';
 import * as fromRoot from '../../../reducers';
 import * as data from '../../actions/data';
 import * as router from '../../actions/router';
 import { MatTableDataSource } from '@angular/material';
 
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/operator/map';
-import { query } from '@angular/animations';
+// import { query } from '@angular/animations';
 
 @Component({
   selector: 'app-data-table',
@@ -34,21 +33,25 @@ export class DataTableComponent implements OnInit {
     this.pageSize$ = this.store.pipe(select(fromRoot.getPageSize));
     this.length$ = this.store.pipe(select(fromRoot.getLength));
     this.columns$ = this.store.pipe(select(fromRoot.getColumns));
-    this.columnNames$ = this.columns$.map(columns => {
-      return columns.map(c => {
-        return c.name;
-      });
-    });
-    this.pageIndex$ = this.store.pipe(select(fromRoot.getStart))
-      .withLatestFrom(this.pageSize$)
-      .map(([start, pageSize]) => {
+    this.columnNames$ = this.columns$.pipe(
+      map(columns => {
+        return columns.map(c => c.name);
+      })
+    );
+    this.pageIndex$ = this.store.pipe(
+      select(fromRoot.getStart),
+      withLatestFrom(this.pageSize$),
+      map(([start, pageSize]) => {
         return start / pageSize;
-      });
+      })
+    );
 
-    this.data$ = this.store.pipe(select(fromRoot.getRows))
-      .map(rows => {
+    this.data$ = this.store.pipe(
+      select(fromRoot.getRows),
+      map(rows => {
         return new MatTableDataSource<Object>(rows);
-      });
+      })
+    );
   }
 
   ngOnInit() {
