@@ -1,13 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Effect, Actions, ofType } from '@ngrx/effects';
-import { Action, Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/observable/fromPromise';
+import { Store } from '@ngrx/store';
+import { from } from 'rxjs';
+import { switchMap, mergeMap, withLatestFrom, map } from 'rxjs/operators';
 
 import { State } from '../../reducers';
 import { Wakanda } from '../../wakanda';
@@ -26,15 +21,11 @@ import * as routerActions from '../actions/router';
 @Injectable()
 export class DataEffects {
     @Effect()
-    fetch$ = this.actions$
-        .ofType<FetchData>(
-            DataActionTypes.FetchData
-        )
-        .switchMap(() => {
-            return this.wakanda.getCatalog();
-        })
-        .withLatestFrom(this.store$)
-        .switchMap(([ds, state]: [any, State]) => {
+    fetch$ = this.actions$.pipe(
+        ofType<FetchData>(DataActionTypes.FetchData),
+        switchMap(() => this.wakanda.getCatalog()),
+        withLatestFrom(this.store$),
+        switchMap(([ds, state]: [any, State]) => {
             let query = state.data.query;
             let tableName = state.data.tableName;
             let pageSize = state.data.pageSize;
@@ -50,44 +41,47 @@ export class DataEffects {
                 pageSize: pageSize,
                 start: start
             });
-        })
-        .map((response: any) => {
+        }),
+        map((response: any) => {
             return new UpdateData({
                 entities: response.entities,
                 length: response._count
             });
-        });
+        })
+    );
 
     @Effect()
-    fetchTables = this.actions$
-        .ofType<FetchTables>(DataActionTypes.FetchTables)
-        .mergeMap(action => {
+    fetchTables = this.actions$.pipe(
+        ofType<FetchTables>(DataActionTypes.FetchTables),
+        mergeMap(action => {
             return new Promise((resolve, reject) => {
                 this.wakanda.getCatalog()
                     .then(c => {
                         resolve([c, action.payload]);
                     });
             })
-        })
-        .map(([catalog, table]) => {
+        }),
+        map(([catalog, table]) => {
             let tables = Object.keys(catalog);
 
             return new UpdateTables(tables);
-        });
+        })
+    );
 
     @Effect()
-    fetchColumns$ = this.actions$
-        .ofType<FetchColumns>(DataActionTypes.FetchColumns)
-        .switchMap(() => {
-            return Observable.fromPromise(this.wakanda.getCatalog());
-        })
-        .withLatestFrom(this.store$)
-        .map(([ds, state]: [any, State]) => {
+    fetchColumns$ = this.actions$.pipe(
+        ofType<FetchColumns>(DataActionTypes.FetchColumns),
+        switchMap(() => {
+            return from(this.wakanda.getCatalog());
+        }),
+        withLatestFrom(this.store$),
+        map(([ds, state]: [any, State]) => {
             let tableName = state.data.tableName;
 
             let columns = ds[state.data.tableName].attributes;
             return new UpdateColumns(columns);
-        });
+        })
+    );
 
     constructor(
         private actions$: Actions,
